pyproject.toml
[tool.poetry]
name = "drf-celery-payments"
version = "0.1.0"
description = "REST service for payout requests with Celery"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "payouts" }, { include = "config" }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14"
celery = "^5.3"
redis = "^5.0"
dj-database-url = "^2.1"
psycopg = { version = "^3.1", extras = ["binary"] }
gunicorn = "^21.2"
drf-spectacular = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-django = "^4.5"
pytest-mock = "^3.12"
ruff = "^0.14.11"
mypy = "^1.19.1"
django-stubs = "^5.2.8"
djangorestframework-stubs = "^3.16.7"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["payouts/migrations"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
warn_unused_ignores = true
no_implicit_optional = true
ignore_missing_imports = true
exclude = "payouts/migrations"

[tool.django-stubs]
django_settings_module = "config.settings"
